+++ examples/simpleswitch/setup.jl
@@ -2,17 +2,7 @@ using QuantumSavory
 using QuantumSavory.ProtocolZoo
 using Graphs
 using ConcurrentSim
-using ResumableFunctions
 using Distributions
-
-"""A helper function to repeatedly make requests from a client to the switch to be connected to another client."""
-@resumable function make_request(sim, net, client, other_client, rate_observable)
-    while true
-        wait_time = rand(Exponential(1/rate_observable[]))
-        @yield timeout(sim, wait_time)
-        put!(channel(net, client=>1), Tag(SwitchRequest(client, other_client)))
-    end
-end
 
 function prepare_simulation()
     n = 5   # number of clients
@@ -32,7 +22,11 @@ function prepare_simulation()
     rate_scale = 1/length(client_pairs)
     rates = [Observable(rate_scale) for _ in client_pairs]
     for ((client1, client2), rate) in zip(client_pairs, rates)
-        @process make_request(sim, net, client1, client2, rate)
+        requester = SwitchRequesterProt(
+            sim, net, 1, client1, client2;
+            request_interval=() -> rand(Exponential(1/rate[])),
+        )
+        @process requester()
     end
 
     # Set up the entanglement trackers at each client